Implement dynamic + static value mixing for babel-plugin-htm
