+++ test/babel.test.mjs
@@ -111,6 +111,45 @@ describe('htm/babel', () => {
 		).toBe(`h("a",Object.assign({b:"1"},foo,{c:2},{d:3}));`);
 	});
 
+	test('mix-and-match dynamic and static values', () => {
+		expect(
+			transform('html`<a b="1${2}${3}"></a>`;', {
+				...options,
+				plugins: [
+					[htmBabelPlugin, {
+						useBuiltIns: true
+					}]
+				]
+			}).code
+		).toBe(`h("a",{b:"1"+2+3});`);
+	});
+
+	test('coerces props to strings when needed', () => {
+		expect(
+			transform('html`<a b=\'${1}${2}${"3"}${4}\'></a>`;', {
+				...options,
+				plugins: [
+					[htmBabelPlugin, {
+						useBuiltIns: true
+					}]
+				]
+			}).code
+		).toBe(`h("a",{b:1+""+2+"3"+4});`);
+	});
+
+	test('coerces props to strings only when needed', () => {
+		expect(
+			transform('html`<a b=\'${"1"}${2}${"3"}${4}\'></a>`;', {
+				...options,
+				plugins: [
+					[htmBabelPlugin, {
+						useBuiltIns: true
+					}]
+				]
+			}).code
+		).toBe(`h("a",{b:"1"+2+"3"+4});`);
+	});
+
 	describe('{variableArity:false}', () => {
 		test('should pass no children as an empty Array', () => {
 			expect(
